Split out TargetPixelFileFactory from TargetPixelFile
